requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
azure-core==1.38.2
    # via azure-storage-blob
azure-functions==1.24.0
    # via fx-pipeline
azure-storage-blob==12.28.0
    # via fx-pipeline
certifi==2026.2.25
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.5
    # via azure-storage-blob
duckdb==1.4.4
    # via fx-pipeline
idna==3.11
    # via requests
isodate==0.7.2
    # via azure-storage-blob
markupsafe==3.0.3
    # via werkzeug
polars==1.38.1
    # via fx-pipeline
polars-runtime-32==1.38.1
    # via polars
pyarrow==23.0.1
    # via fx-pipeline
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
requests==2.32.5
    # via
    #   azure-core
    #   fx-pipeline
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-storage-blob
urllib3==2.6.3
    # via requests
werkzeug==3.1.6
    # via azure-functions
